+++ packages/react/src/components/index.tsx
@@ -1,3 +1,4 @@
 export * from './modal';
 export * from './tab';
-export * from './popover1';
+export * from './popover';
+export * from  './carousel'
