+++ packages/react/src/index.ts
@@ -1,4 +1,3 @@
 export * from './components'
 export * from './elements'
 export * from './layout';
-export  * from './components/carousel'
